Add files via upload
This is JavaScript only, in the near future I will be integrating this with HTML and CSS making it into an interactive site. I'll also change the code when I learn a simpler way to do it, at present this was coded using conditionals. The code generates a random number between 0 and 200 and the answers that the 8-Ball gives are determined on which number the code randomly spits out!